node: move the mining job store out of the rpc layer
job_store holds the transaction selection behind a getblocktemplatelight job so
submitblocklight can rebuild the block from the solved header and coinbase. It
lived in kth::node::rpc and compiled only with KTH_WITH_RPC, but it is a plain
mining helper with no RPC dependency, and the SV2 template provider will need
the same id-to-selection store.

Move it to kth::node::mining (src/node/{include,src}/mining/job_store.hpp/.cpp),
built as part of the node core regardless of the RPC option. The RPC
method_context and server now reference mining::job_store. No behavior change.

The unit test moves with it (mining_job_store.cpp, tagged [mining job_store]) and
out of the RPC-gated test block, so it runs in the default build too.
