+++ src/node/test/mining_job_store.cpp
@@ -8,10 +8,10 @@
 #include <vector>
 
 #include <kth/domain.hpp>
-#include <kth/node/rpc/job_store.hpp>
+#include <kth/node/mining/job_store.hpp>
 
 using namespace kth;
-using namespace kth::node::rpc;
+using namespace kth::node::mining;
 
 namespace {
 
@@ -31,9 +31,9 @@ std::vector<transaction_const_ptr> txs(std::initializer_list<uint32_t> locktimes
 
 } // namespace
 
-// Start Test Suite: rpc job_store tests
+// Start Test Suite: mining job_store tests
 
-TEST_CASE("job_store add then get roundtrips", "[rpc job_store]") {
+TEST_CASE("job_store add then get roundtrips", "[mining job_store]") {
     job_store store(10, 3600);
     auto const id = store.add(txs({1, 2}));
     auto const got = store.get(id);
@@ -41,24 +41,24 @@ TEST_CASE("job_store add then get roundtrips", "[rpc job_store]") {
     REQUIRE(got->size() == 2u);
 }
 
-TEST_CASE("job_store add is idempotent for equal selections", "[rpc job_store]") {
+TEST_CASE("job_store add is idempotent for equal selections", "[mining job_store]") {
     job_store store(10, 3600);
     auto const a = store.add(txs({7, 8, 9}));
     auto const b = store.add(txs({7, 8, 9}));
     REQUIRE(a == b);
 }
 
-TEST_CASE("job_store distinct selections get distinct ids", "[rpc job_store]") {
+TEST_CASE("job_store distinct selections get distinct ids", "[mining job_store]") {
     job_store store(10, 3600);
     REQUIRE(store.add(txs({1})) != store.add(txs({2})));
 }
 
-TEST_CASE("job_store returns nullopt for unknown id", "[rpc job_store]") {
+TEST_CASE("job_store returns nullopt for unknown id", "[mining job_store]") {
     job_store store(10, 3600);
     REQUIRE_FALSE(store.get("deadbeef").has_value());
 }
 
-TEST_CASE("job_store evicts the oldest past the count bound", "[rpc job_store]") {
+TEST_CASE("job_store evicts the oldest past the count bound", "[mining job_store]") {
     job_store store(/*max_jobs*/ 2, /*ttl*/ 3600);
     auto const first = store.add(txs({1}));
     store.add(txs({2}));
